+++ sram.h
@@ -6,8 +6,8 @@
 
 void initSRAM(void);
 
-uint8_t readSRAM(uint8_t address);
-void writeSRAM(uint8_t address, uint8_t data);
+uint8_t readSRAM(uint16_t address);
+void writeSRAM(uint16_t address, uint8_t data);
 
 	
 void SRAM_test(void);
